requirements.txt
python-telegram-bot==20.3
gspread==5.7.2
google-auth==2.3.0
python-dotenv==1.0.0  # (Jika pakai .env)